requirements.txt
# Core PyTorch dependencies
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Deep learning utilities
torchsummary==1.5.1
timm==1.0.9
torchmetrics==1.4.1

# Data processing and visualization
numpy==1.26.4
matplotlib==3.9.2
Pillow==10.4.0
tqdm==4.66.5

# Image augmentation
imgaug==0.4.0
opencv-python==4.10.0.84

# Experiment tracking
comet_ml==3.45.0

# Custom dataset
LymphoMNIST==0.0.1

# Utilities
dataclasses==0.6
